🌦 fix weather
Switch to open weather map instead of yahoo, since the yahoo open
weather api has been offlined as of January 3, 2019.

Add a default api token, and a settings field for it in case the user
needs to insert their own.

Fix #3, resize date field when date length is > 21 characters to prevent
cropping.